code/prometheus-k8s-operator/pyproject.toml
# Copyright 2021 Canonical Ltd.
# See LICENSE file for licensing details.
[project]
name = "prometheus-k8s"
version = "0.0"
requires-python = ">=3.13"

dependencies = [
  "ops",
  "pyyaml",
  "cosl",
  "lightkube>=0.11",
  "lightkube-models>=1.22.0.4",
  "cryptography",
  "jsonschema",
  "requests",
  "pydantic>=2",
]

[project.optional-dependencies]
dev = [
  # Linting
  "ruff",
  "codespell",
  # Static
  "pyright",
  # Unit
  "pytest",
  "coverage[toml]",
  "ops[testing]",
  "responses",
  "setuptools",
]

# Testing tools configuration
[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

# Linting tools configuration
[tool.ruff]
line-length = 99
extend-exclude = ["__pycache__", "*.egg_info"]

[tool.ruff.lint]
select = ["E", "W", "F", "C", "N", "R", "D", "I001", "B018"]
# Ignore E501 because using black creates errors with this
# Ignore D107 Missing docstring in __init__
ignore = ["F403", "F405", "E501", "D107", "RET504", "C901", "D100", "D101", "D102", "D103", "D104"]
# D100, D101, D102, D103: Ignore missing docstrings in tests
per-file-ignores = {"tests/*" = ["D100","D101","D102","D103"]}

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
venvPath = "."
venv = ".venv"
extraPaths = ["src"]
pythonVersion = "3.13"
pythonPlatform = "All"
reportWildcardImportFromLibrary = "none"
reportUnusedExpression = "error"

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"
addopts = "--tb=native --verbose --capture=no --log-cli-level=INFO"

[tool.codespell]
skip = ".git,.tox,build,venv*"
ignore-words-list = "assertIn"
